split requestium into multiple files
