+++ .gitignore
@@ -100,6 +100,9 @@ ENV/
 # mkdocs documentation
 /site
 
+# Poetry
+poetry.lock
+
 # mypy
 .mypy_cache/
