+++ requestium/__init__.py
@@ -5,4 +5,4 @@ from selenium.webdriver.common.by import By  # noqa: F401
 from selenium.webdriver.common.keys import Keys  # noqa: F401
 from selenium.webdriver.support.ui import Select  # noqa: F401
 
-from .requestium import Session  # noqa: F401
+from .requestium_session import Session  # noqa: F401
